Split linux check
